Reorganize test files
Reorganize the test files to more closely
reflect the non-test file structure.

- Moved test_communication.py -> test_connection.py
- Moved test_ens_on_host.py -> test_connection.py
- Moved test_model.py -> test_simulator.py
- Moved test_multiple.py -> test_node.py
- Moved test_precompute.py -> test_simulator.py
- Moved test_radius.py -> test_ensemble.py
- Moved test_snips.py -> test_simulator.py
- Moved emulator/tests/test_emulator.py ->
  emulator/tests/test_interface.py
- Moved hardware/tests/test_simulator.py ->
  hardware/tests/test_interface.py
- Moved test_simulator.py::test_snip_input_count ->
  hardware/tests/test_interface.py
