+++ nengo_loihi/emulator/tests/test_interface.py
