+++ nengo_loihi/hardware/tests/test_interface.py
@@ -117,3 +117,16 @@ def test_interface_connection_errors(Simulator):
 
         with pytest.raises(SimulationError, match="[Cc]ould not connect"):
             interface.connect(attempts=1)
+
+
+@pytest.mark.skipif(pytest.config.getoption("--target") != "loihi",
+                    reason="snips only exist on loihi")
+def test_snip_input_count(Simulator, seed, plt):
+    with nengo.Network(seed=seed) as model:
+        a = nengo.Ensemble(100, 1)
+        for i in range(30):
+            stim = nengo.Node(0.5)
+            nengo.Connection(stim, a, synapse=None)
+    with Simulator(model) as sim:
+        with pytest.warns(UserWarning, match="Too many spikes"):
+            sim.run(0.01)
