+++ nengo_loihi/tests/test_inputs.py
@@ -2,7 +2,6 @@ import nengo
 from nengo.exceptions import SimulationError
 import pytest
 
-
 from nengo_loihi.inputs import ChipReceiveNode
