+++ nengo_loihi/tests/test_node.py
@@ -1,5 +1,5 @@
-import nengo
 import numpy as np
+import nengo
 import pytest
